fix iax and add regex dialplan
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@382 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/iax/src/iax.c
@@ -66,7 +66,7 @@ void gettimeofday(struct timeval *tv, void /*struct timezone*/ *tz);
 #include "jitterbuf.h"
 #endif
 
-typedef int64_t time_in_ms_t;
+typedef long long time_in_ms_t;
 
 #include "iax-client.h"
 #include "md5.h"
